refactor(dispatcher): extract fmtAptShort/fmtApt to shared date-fmt util (FU-103)
Resolves: FU-103
