+++ components/dispatcher/planner/MoveCardCompact.jsx
@@ -1,21 +1,7 @@
 import { memo } from 'react';
 import { getLoadTypeColor } from '../../../lib/dispatcher/load-type-colors';
 import { lfdPillClass, fmtLfdShort } from '../../../lib/dispatcher/lfd-urgency';
-
-function fmtAptShort(iso) {
-  if (!iso) return null;
-  try {
-    const d = new Date(iso);
-    if (Number.isNaN(d.getTime())) return null;
-    const mm = String(d.getMonth() + 1).padStart(2, '0');
-    const dd = String(d.getDate()).padStart(2, '0');
-    const hh = String(d.getHours()).padStart(2, '0');
-    const mi = String(d.getMinutes()).padStart(2, '0');
-    return `${mm}/${dd} ${hh}:${mi}`;
-  } catch {
-    return null;
-  }
-}
+import { fmtAptShort } from '../../../lib/dispatcher/date-fmt';
 
 /**
  * Compact card view used by both the right-rail unassigned panel and the
